+++ src/frontends/xim/src/main.rs
@@ -10,6 +10,13 @@ mod pe_window;
 fn main() -> Result<(), ServerError> {
     let mut args = pico_args::Arguments::from_env();
 
+    if args.contains(["-h", "--help"]) {
+        println!("-h or --help: show help");
+        println!("-v or --version: show version");
+        println!("--verbose: more verbose log");
+        return Ok(());
+    }
+
     if args.contains(["-v", "--version"]) {
         kime_version::print_version!();
         return Ok(());
@@ -21,7 +28,7 @@ fn main() -> Result<(), ServerError> {
         log::LevelFilter::Info
     };
 
-    if args.contains(["-l", "--log"]) {
+    if args.contains("--verbose") {
         log_level = log::LevelFilter::Trace;
     }