+++ src/frontends/wayland/src/main.rs
@@ -232,9 +232,15 @@ impl KimeContext {
 fn main() {
     let mut args = pico_args::Arguments::from_env();
 
-    if args.contains("--version") {
-        println!("{}: {}", env!("CARGO_PKG_NAME"), env!("CARGO_PKG_VERSION"));
+    if args.contains(["-h", "--help"]) {
+        println!("-h or --help: show help");
+        println!("-v or --version: show version");
+        println!("--verbose: more verbose log");
+        return;
+    }
 
+    if args.contains(["-v", "--version"]) {
+        kime_version::print_version!();
         return;
     }
 
@@ -251,7 +257,7 @@ fn main() {
         log::LevelFilter::Info
     };
 
-    if args.contains("--log") {
+    if args.contains("--verbose") {
         log_level = log::LevelFilter::Trace;
     }