+++ docs/CHANGELOG.md
@@ -3,6 +3,7 @@
 * Show hangul/english state on tray icon
 * Set gtk log domain to `kime`
 * Let kime-wayland exit when IO Error occured
+* Fix wayland input bug
 
 # 0.9.1